Do not round mirror placement
Rounding was initially added to work around an issue in chrome that
caused blurriness when fractional values were used in translate3d().
However, since then, these bugs appear to have been fixed.

Rounding causes the mirror to not appear in the exact position that the
original object was. Since the values here are CSS pixels and not real
pixels, on some devices this rounding can be more than one real pixel.